deps: upgrade v8 to 4.2.77.13
This commit applies some secondary changes in order to make `make test`
pass cleanly:

* disable broken postmortem debugging in common.gypi

* drop obsolete strict mode test in parallel/test-repl

* drop obsolete test parallel/test-v8-features

PR-URL: https://github.com/iojs/io.js/pull/1232

+++ deps/v8/src/unicode-decoder.cc
@@ -10,16 +10,16 @@
 
 namespace unibrow {
 
-void Utf8DecoderBase::Reset(uint16_t* buffer, unsigned buffer_length,
-                            const uint8_t* stream, unsigned stream_length) {
+void Utf8DecoderBase::Reset(uint16_t* buffer, size_t buffer_length,
+                            const uint8_t* stream, size_t stream_length) {
   // Assume everything will fit in the buffer and stream won't be needed.
   last_byte_of_buffer_unused_ = false;
   unbuffered_start_ = NULL;
   bool writing_to_buffer = true;
   // Loop until stream is read, writing to buffer as long as buffer has space.
-  unsigned utf16_length = 0;
+  size_t utf16_length = 0;
   while (stream_length != 0) {
-    unsigned cursor = 0;
+    size_t cursor = 0;
     uint32_t character = Utf8::ValueOf(stream, stream_length, &cursor);
     DCHECK(cursor > 0 && cursor <= stream_length);
     stream += cursor;
@@ -56,9 +56,9 @@ void Utf8DecoderBase::Reset(uint16_t* buffer, unsigned buffer_length,
 
 
 void Utf8DecoderBase::WriteUtf16Slow(const uint8_t* stream, uint16_t* data,
-                                     unsigned data_length) {
+                                     size_t data_length) {
   while (data_length != 0) {
-    unsigned cursor = 0;
+    size_t cursor = 0;
     uint32_t character = Utf8::ValueOf(stream, Utf8::kMaxEncodedSize, &cursor);
     // There's a total lack of bounds checking for stream
     // as it was already done in Reset.
